Sending/Parsing peer messages implemented + tests.
Can now construct a Message and send it over a connection. Similarly,
we can take a byte array and produce a Message if it is able to be
parsed. Added some simple tests to verify that things are being parsed
correctly. I think everything is ready to start torrenting :)
Besides file stuff of course... but I think I'll implement that in the
main package instead of in the client.