Framework for connecting to peers in place.
I think the system for connecting to peers is correct... I'm a little
worried that ignoring safe exits in this program is shortsighted and
will cause problems in the future, but I'm deciding to forego that as
this is a simple exercise and a command line program. Next up is
putting keep alive messages in.